// accompany-consullting/src/app/Model/recuperation.ts
export class Recuperation {
    demandeur: number;
    validateur: number;
    dateDebut: Date;
    dateFin: Date;
    type: string;
    etat: string;
  id: number;

    duree : number ;
    commentaire:string;
    validateurnom :string;
    validateurmail : string ;
   demandeurnom :string;
   numeroDemande : number;
   manager : string ;
   joursValides : any;
   showInput: boolean= false ;
   show : boolean = false ;
}